Address review: isolate the daemon probe into its own testable service; fix default-config query
Second autoreview round on the daemon probe (#6629):

- [P2] When configDir is nil the probe kept the app's inherited CLAUDE_CONFIG_DIR,
  querying a different daemon than the default/account roots the index validates
  transcripts against (and than the "" cache key). Clear CLAUDE_CONFIG_DIR when no
  trusted config is present so the queried daemon matches the validated roots.

- [P2, Aziz] The subprocess/cache/lock/singleton logic was a separate responsibility
  bloating the already-oversized RestorableAgentSession.swift and forced a file-length
  budget bump. Extract it to Sources/ClaudeBackgroundAgentsQuery.swift behind injected
  `probe` and `now` seams (defaulting to the real `claude agents --json` runner and the
  system clock), leaving RestorableAgentSessionIndex depending only on the provider
  closure. RestorableAgentSession.swift is now ~117 lines over its prior size (under the
  250-line threshold).

Adds unit tests for the cache/TTL/per-config-dir behaviour and cachedOnly's TTL using
the injected seams (no subprocess). Wires the new file into the Xcode project.
