+++ Sources/RestorableAgentSession.swift
@@ -1,4 +1,3 @@
-import Darwin
 import Foundation
 import CMUXAgentLaunch
 
@@ -2059,165 +2058,5 @@ private extension CmuxTopProcessArguments {
             return nil
         }
         return UUID(uuidString: rawValue)
-    }
-}
-
-/// Queries the Claude Code daemon (`claude agents --json`) for live background agents so a
-/// transcript-less ghost panel can be reconciled to its real session id.
-/// https://github.com/manaflow-ai/cmux/issues/6622
-///
-/// Best-effort and self-contained: any failure (claude not found, non-zero exit, malformed
-/// JSON, or the bounded probe timing out) yields no agents, so the restorable index degrades
-/// to its prior behaviour rather than blocking or surfacing a wrong session. Results are
-/// cached per `CLAUDE_CONFIG_DIR` for a short TTL so the event-driven, off-main index reload
-/// does not respawn `claude` on every refresh. `@unchecked Sendable`: the cache is guarded by
-/// a lock. Use ``live(configDir:)`` from the off-main loaders (it may spawn the probe) and
-/// ``cachedOnly(configDir:)`` from synchronous main-thread save paths (it only reuses the
-/// cache the off-main loaders populate, never spawning a subprocess).
-final class ClaudeBackgroundAgentsQuery: @unchecked Sendable {
-    static let shared = ClaudeBackgroundAgentsQuery()
-
-    private let lock = NSLock()
-    // Serializes the probe so a cache miss observed by two concurrent off-main loaders does
-    // not spawn `claude agents --json` twice; the post-acquire re-check returns the first
-    // fetch's freshly cached result to the second caller.
-    private let fetchLock = NSLock()
-    private var cache: [String: (agents: [ClaudeBackgroundAgentSnapshot], fetchedAt: Date)] = [:]
-    private static let cacheTTL: TimeInterval = 20
-    private static let probeTimeout: TimeInterval = 5
-    private static let terminateGrace: TimeInterval = 1
-
-    /// Off-main: returns cached agents within the TTL, otherwise runs the bounded probe and
-    /// caches the result. Must not be called on the main thread.
-    func live(configDir: String?) -> [ClaudeBackgroundAgentSnapshot] {
-        let key = configDir ?? ""
-        if let fresh = freshCachedAgents(forKey: key) {
-            return fresh
-        }
-
-        fetchLock.lock()
-        defer { fetchLock.unlock() }
-        // Re-check under the fetch lock: a concurrent probe for this key may have just filled
-        // the cache while we waited.
-        if let fresh = freshCachedAgents(forKey: key) {
-            return fresh
-        }
-
-        let agents = Self.fetch(configDir: configDir)
-        lock.lock()
-        cache[key] = (agents, Date())
-        lock.unlock()
-        return agents
-    }
-
-    private func freshCachedAgents(forKey key: String) -> [ClaudeBackgroundAgentSnapshot]? {
-        lock.lock()
-        defer { lock.unlock() }
-        guard let cached = cache[key], Date().timeIntervalSince(cached.fetchedAt) < Self.cacheTTL else {
-            return nil
-        }
-        return cached.agents
-    }
-
-    /// Main-thread safe: returns whatever the off-main loaders have already cached for this
-    /// config dir, never spawning `claude`. The synchronous session-save path uses this so it
-    /// reconciles from the warm cache without a subprocess on the main thread. Applies the same
-    /// TTL as ``live(configDir:)`` so a stale entry — where the daemon would now report zero, a
-    /// different, or multiple ambiguous agents — degrades to no reconciliation instead of
-    /// persisting a wrong session.
-    func cachedOnly(configDir: String?) -> [ClaudeBackgroundAgentSnapshot] {
-        freshCachedAgents(forKey: configDir ?? "") ?? []
-    }
-
-    /// Holds the non-`Sendable` `Process`/`FileHandle` for use inside the background read and
-    /// the watchdog without tripping strict-concurrency capture checks; the objects are only
-    /// touched from this single probe.
-    private final class ProbeBox: @unchecked Sendable {
-        let process: Process
-        let readHandle: FileHandle
-        init(process: Process, readHandle: FileHandle) {
-            self.process = process
-            self.readHandle = readHandle
-        }
-    }
-
-    private final class DataBox: @unchecked Sendable {
-        private let lock = NSLock()
-        private var value = Data()
-        func set(_ data: Data) { lock.lock(); value = data; lock.unlock() }
-        func get() -> Data { lock.lock(); defer { lock.unlock() }; return value }
-    }
-
-    private static func fetch(configDir: String?) -> [ClaudeBackgroundAgentSnapshot] {
-        guard let plan = resolvedClaudeLaunchPlan() else { return [] }
-
-        let process = Process()
-        process.executableURL = plan.executableURL
-        process.arguments = ["agents", "--json"]
-        // Use the resolver's launch environment (which rebuilds PATH to include Homebrew,
-        // ~/.local/bin, nvm/mise/asdf, etc.), overlaying CLAUDE_CONFIG_DIR. A Dock-launched
-        // macOS app's own PATH is stripped, so running the resolved claude (often a shim that
-        // execs `/usr/bin/env node`) under it would exit nonzero and silently no-op the probe.
-        var environment = plan.environment
-        if let configDir, !configDir.isEmpty {
-            environment["CLAUDE_CONFIG_DIR"] = (configDir as NSString).expandingTildeInPath
-        }
-        process.environment = environment
-
-        let stdout = Pipe()
-        process.standardOutput = stdout
-        process.standardError = FileHandle.nullDevice
-        process.standardInput = FileHandle.nullDevice
-
-        // Signal on real process exit so the exit wait below is bounded too, not just the read.
-        let exited = DispatchSemaphore(value: 0)
-        process.terminationHandler = { _ in exited.signal() }
-
-        do {
-            try process.run()
-        } catch {
-            return []
-        }
-
-        // Drain stdout on a background queue so the probe stays bounded even if the read would
-        // otherwise block: each wait below has a deadline and escalates SIGTERM -> SIGKILL on
-        // expiry, which closes stdout and forces the process to exit.
-        let box = ProbeBox(process: process, readHandle: stdout.fileHandleForReading)
-        let output = DataBox()
-        let readDone = DispatchSemaphore(value: 0)
-        DispatchQueue.global(qos: .utility).async {
-            output.set(box.readHandle.readDataToEndOfFile())
-            readDone.signal()
-        }
-
-        func terminateAndDrain() {
-            if box.process.isRunning { box.process.terminate() }
-            if exited.wait(timeout: .now() + Self.terminateGrace) == .timedOut {
-                if box.process.isRunning { kill(box.process.processIdentifier, SIGKILL) }
-                _ = exited.wait(timeout: .now() + Self.terminateGrace)
-            }
-            _ = readDone.wait(timeout: .now() + Self.terminateGrace)
-        }
-
-        // Bound the read.
-        if readDone.wait(timeout: .now() + Self.probeTimeout) == .timedOut {
-            terminateAndDrain()
-            return []
-        }
-        // Bound process exit: stdout can close while the process hangs, so do not call the
-        // unbounded `waitUntilExit()` — wait on the termination handler with a deadline.
-        if exited.wait(timeout: .now() + Self.terminateGrace) == .timedOut {
-            terminateAndDrain()
-            return []
-        }
-
-        guard process.terminationStatus == 0 else { return [] }
-        return ClaudeBackgroundAgentReconciler.parse(agentsJSON: output.get())
-    }
-
-    private static func resolvedClaudeLaunchPlan() -> AgentSessionLaunchPlan? {
-        try? AgentExecutableResolver(
-            configuredExecutablePaths: AgentExecutableResolver.cmuxConfiguredExecutablePaths()
-        ).resolve(.claude)
     }
 }
